+++ release-notes.md
@@ -1,5 +1,5 @@
 ## @thu-info/app
-- 修复了某些情况下 APP 功能异常的问题
+- 修复了电子发票无法查看的问题
 
 ## @thu-info/lib
-- fix: Redirect to WebVPN via oauth. to solve in-campus CSRF error
+- fix: fixed `helper.getInvoicePDF`